Chore: sync Rhiza template v0.18.8 → v0.19.3 (#57)
* Chore: bump rhiza template ref v0.18.8 → v0.19.3


* Chore: sync Rhiza template v0.18.8 → v0.19.3

Resolve the 3-way merge from the template bump. Conflict resolution:

- Rhiza-managed workflow caller stubs (rhiza_benchmark/book/ci/codeql/marimo/
  release/sync/weekly): took upstream — bumped pins to v0.19.3 and folded in
  the new least-privilege per-job permissions blocks, the CI Python-matrix
  source-of-truth comment, and the sync top-level read-only permissions.
- rhiza_release.yml: dropped the redundant update-changelog job, added build/
  finalise-release permissions, and staged the SLSA provenance bundle as a
  release artifact (all upstream).
- New managed files added by the bump: rhiza_fuzzing/mutation/scorecard
  workflows, .github/CONFIG.md, discussion templates, cliff.toml, the
  .rhiza/tests/utils tests, and the rhiza_book Claude command.
- SECURITY.md and PULL_REQUEST_TEMPLATE.md: applied upstream edits (SBOM
  attestation claim; CHANGELOG checklist item).
- .rhiza/.env: added upstream descriptive header, kept local
  MARIMO_FOLDER=book/marimo/notebooks override.
- ruff.toml (locally hardened): merged upstream's stricter rule additions
  (A/ANN/ARG/BLE/PIE + rationale block) while keeping local intent.
- pytest.ini (locally hardened): kept as-is; upstream's log_cli=false intent
  is already satisfied (no live logging configured locally).


* Fix: satisfy v0.19.3 tightened lint/type/docs gates

The template bump enabled stricter gates that surfaced pre-existing issues in
locally-owned code:

- norm.py: ruff A002 (flake8-builtins) flags the `ord` parameter as shadowing
  a builtin; it intentionally mirrors np.linalg.norm's public API, so suppress
  with a targeted noqa rather than renaming (breaking change).
- lstsq.py: mypy --strict (new cross-check) rejects the np.linalg.lstsq return
  whose static dtype is floating[Any]/signedinteger; coerce to the declared
  float64/int public contract (values are already float64 at runtime).
- test_svd.py: expanded docs-coverage scope requires a docstring on the nested
  fake_svd helper.


---------

+++ src/cvx/linalg/lstsq.py
@@ -84,4 +84,9 @@ def lstsq(
 
     _warn_ill_conditioned(cond, cond_threshold)
 
-    return x, residuals, rank, sv
+    return (
+        x.astype(np.float64, copy=False),
+        residuals.astype(np.float64, copy=False),
+        int(rank),
+        sv.astype(np.float64, copy=False),
+    )
